refactor: introduce wrapper + message_pack_format top-level modules
Move PR #36–#45 wire-format-aligned variants out of src/sketches/ into a
new src/wrapper/ module so the in-process algorithms and the
sketchlib-go-aligned producer types live in separate trees. Add a
src/message_pack_format.rs placeholder noting that asap_sketchlib (Rust)
and sketchlib-go each maintain their own MessagePack representation,
byte-compatible at the envelope level.

Files moved:
- src/wrapper/{ddsketch,countminsketch,countsketch,hll,kll,countminsketch_topk}.rs
  carry the wire-format types appended after the per-file banner in PR #36+.
- src/wrapper/{hydra_kll,set_aggregator,delta_set_aggregator}.rs are
  whole-file moves (they were entirely wire-format).

CountL2HH stays in src/sketches/countsketch_topk.rs — it's an in-process
top-k algorithm used by sketch_framework, not a wire-format type.

No semantic changes; cargo test --lib --all-features still 447 passed.
